requirements.txt
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Django==3.0.5
djangorestframework==3.12.2
djongo==1.3.3
dnspython==2.1.0
firebase-admin==4.5.1
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
google-crc32c==1.1.1
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.1
httplib2==0.18.1
idna==2.10
msgpack==1.0.2
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.2
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.2
